requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
audioop-lts==0.2.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
comtypes==1.4.10
distro==1.9.0
Django==5.2.2
django-cors-headers==4.7.0
django-jazzmin==3.0.1
djangorestframework==3.16.0
docopt==0.6.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
Js2Py==0.74
openai==1.70.0
packaging==24.2
pillow==11.2.1
pipwin==0.5.2
PyAudio==0.2.14
pydantic==2.11.1
pydantic_core==2.33.0
pyjsparser==2.7.1
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
pyttsx3==2.98
pywin32==310
RapidFuzz==3.13.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.14.2
sqlparse==0.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
